requirements.txt
dj-database-url==0.4.1
Django==1.11.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
django-unixdatetimefield==0.1.6
djangorestless==0.0.10
neo4j-driver==1.5.0